+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changelog for NeoFS Node
 - Size-split object is now available even when LINK is unavailable in EC container (#3754)
 - `neofs-adm balance container-status` command (#3693)
 - Container IDs to JSON output of ADM `fschain dump-containers` command (#3789)
+- Support for IDs from JSON to ADM `fschain restore-containers` command (#3789)
 
 ### Fixed
 - IR panics at graceful shutdown (#3706)